+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Retour Changelog
 
+## 3.2.26 - 2025.04.22
+### Added
+* Added the **Resolve Craft Sites** setting to allow user control over whether Craft sites should be factored in to destination URLs ([#334](https://github.com/nystudio107/craft-retour/issues/334))
+* Added validation on Short Link fields so you cannot create one if the Short Link already exists at that URL ([#340](https://github.com/nystudio107/craft-retour/issues/340))
+
 ## 3.2.25 - 2024.04.03
 ### Changed
 * Only look at adding site prefixes to `pathonly` redirects via GraphQL